+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed 
 - Fixed guide description in Recently Viewer [#322](https://github.com/rokwire/illinois-app/issues/322).
+- Do not show favorite button in Student Guide items if privacy level does not support this.
 
 ## [2.4.27] - 2021-06-21
 ### Changed